# docs/yaml/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Hotel Booking API
  description: API for booking hotels and managing accommodations
  contact:
    name: Tso Wan Lung
    email: 247029594@stu.vtc.edu.hk
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "http://localhost:10888/api/v1"
    description: Development server
tags:
  - name: Hotels
    description: Hotel search and information
  - name: Bookings
    description: Booking management
  - name: Users
    description: User account management
  - name: Authentication
    description: Login and authentication

paths:
  /hotels:
    get:
      tags:
        - Hotels
      summary: Get all hotels
      description: Returns a list of all hotels with pagination
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Maximum number of hotels to return
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
      responses:
        "200":
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"
        "404":
          description: No hotels found

  /hotels/search:
    post:
      tags:
        - Hotels
      summary: Search hotels
      description: Search hotels by country, city, and dates
      requestBody:
        description: Search criteria
        required: true
        content:
          application/json:
            schema:
              $ref: "./searchHotels.json#/definitions/searchHotels"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"
        "404":
          description: No hotels found matching the criteria

  /hotels/{id}:
    get:
      tags:
        - Hotels
      summary: Get hotel by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Hotel ID
      responses:
        "200":
          description: Hotel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
        "404":
          description: Hotel not found

  /hotels/{id}/rooms:
    get:
      tags:
        - Hotels
      summary: Get rooms by hotel ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Hotel ID
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "404":
          description: No rooms found for this hotel

  /hotels/bookings:
    post:
      tags:
        - Bookings
      summary: Create a booking
      security:
        - bearerAuth: []
      requestBody:
        description: Booking details
        required: true
        content:
          application/json:
            schema:
              $ref: "./createBooking.json#/definitions/createBooking"
      responses:
        "201":
          description: Booking created successfully
        "400":
          description: Invalid request data
        "401":
          description: Unauthorized
        "500":
          description: Failed to create booking

  /hotels/private/bookings:
    get:
      tags:
        - Bookings
      summary: Get bookings by user role
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: No bookings found

  /hotels/update/bookings:
    post:
      tags:
        - Bookings
      summary: Update a booking
      description: Updates the booking details, room statuses, and adds a message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./updateBooking.json#/definitions/updateBooking"
      responses:
        200:
          description: Booking updated successfully
        400:
          description: Validation failed
        500:
          description: Internal server error

components:
  schemas:
    Hotel:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the hotel
        name:
          type: string
          description: Name of the hotel
        description:
          type: string
          description: Description of the hotel
        city:
          type: string
          description: City where the hotel is located
        country:
          type: string
          description: Country where the hotel is located
        address:
          type: string
          description: Full address of the hotel
        rating:
          type: number
          format: float
          description: Rating of the hotel (0-5)
        review_count:
          type: integer
          description: Number of reviews
        image_url:
          type: string
          description: URL of the hotel image
        links:
          type: object
          properties:
            self:
              type: string
              description: Link to hotel details

    Room:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the room
        hotel_id:
          type: integer
          description: ID of the hotel this room belongs to
        capacity:
          type: integer
          description: Maximum number of guests
        bed_option:
          type: string
          description: Bed configuration (e.g., "1 King Bed")
        amenities:
          type: array
          items:
            type: string
          description: List of amenities
        price_per_night:
          type: number
          format: float
          description: Standard price per night
        has_discount:
          type: boolean
          description: Whether the room has a discount
        discount_rate:
          type: number
          format: float
          description: Discount percentage (0.0-1.0)
        actual_price:
          type: number
          format: float
          description: Price after discount

    Booking:
      type: object
      properties:
        booking_id:
          type: integer
          description: Unique identifier for the booking
        user_id:
          type: integer
          description: ID of the user who made the booking
        start_date:
          type: string
          format: date
          description: Check-in date
        end_date:
          type: string
          format: date
          description: Check-out date
        staff_email:
          type: string
          nullable: true
          description: Email of the assigned staff member
        first_message:
          type: string
          description: Initial message from the user
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
          description: List of rooms in the booking

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT